Update evaluator objects
